More pointers, arrays and strings

+++ 0x06-pointers_arrays_strings/101-print_number.c
@@ -11,7 +11,7 @@ void print_number(int n)
 	if (n < 0)
 	{
 		_putchar('-');
-		n = 0 - n;
+		n = 0 - (unsigned int) n;
 	}
 	if (n >= 10)
 	{
